Small modifications to digest
Cleanup Cargo.toml

Tokio has no feature `main`
Fully specify up to major version

Decompose main.rs
