+++ digest/src/main.rs
@@ -1,10 +1,8 @@
-// use async_openai::config::OpenAIConfig;
-use async_openai::{
-    types::ChatCompletionRequestSystemMessageArgs, types::CreateChatCompletionRequestArgs, Client,
-};
-use chrono::{DateTime, Duration, Utc};
+mod models;
+
+use chrono::Utc;
+use models::{Newsletter, NewsletterConfig};
 use regex::Regex;
-use serde_derive::Deserialize;
 use std::error::Error;
 use std::fs::File;
 use std::io::prelude::*;
@@ -82,6 +80,7 @@ fn create_newsletter_html(
     Ok((newsletter, newsletter_text))
 }
 
+#[allow(clippy::unused_io_amount)]
 fn write_newsletter_to_file(
     newsletter: &mut Newsletter,
     newsletter_text: String,
@@ -92,168 +91,7 @@ fn write_newsletter_to_file(
     }
 }
 
-#[derive(Debug, Deserialize)]
-struct NewsletterConfig {
-    title: String,
-    output_file: String,
-    feeds: Vec<FeedConfig>,
-}
-
-#[derive(Clone, Debug, Deserialize)]
-struct FeedConfig {
-    name: String,
-    feed_url: String,
-    regex_filter: String,
-}
-
-#[derive(Debug)]
-struct Newsletter {
-    title: String,
-    output_file: File,
-    feeds: Vec<Feed>,
-}
-
-impl Newsletter {
-    fn from(config: NewsletterConfig) -> Self {
-        let feeds = config
-            .feeds
-            .iter()
-            .cloned()
-            .map(|feed_config| {
-                let mut feed = Feed::from(feed_config);
-                println!("Open feed {}", feed.feed_url);
-
-                {
-                    feed.fetch_posts().unwrap()
-                };
-                println!("Fetched feed posts");
-                feed
-            })
-            .collect();
-
-        let output_file = File::create(config.output_file).unwrap();
-        println!("Open output file");
-        Self {
-            title: config.title,
-            output_file,
-            feeds,
-        }
-    }
-
-    fn to_html(&mut self) -> String {
-        let mut newsletter_text = String::new();
-
-        for feed in self.feeds.iter_mut() {
-            let feed_title = format!("<h3>{}</h3>", feed.name);
-            newsletter_text.push_str(&feed_title);
-
-            for post in feed.posts.iter_mut() {
-                let post_text = post.to_html();
-
-                newsletter_text.push_str(&post_text);
-            }
-        }
-
-        newsletter_text
-    }
-}
-#[derive(Debug)]
-struct Feed {
-    name: String,
-    feed_url: String,
-    regex_filter: Regex,
-    posts: Vec<Post>,
-}
-
-impl Feed {
-    fn from(config: FeedConfig) -> Self {
-        let regex_filter = Regex::new(&config.regex_filter).unwrap();
-        Self {
-            name: config.name,
-            feed_url: config.feed_url,
-            regex_filter,
-            posts: vec![],
-        }
-    }
-
-    fn fetch_posts(&mut self) -> Result<(), Box<dyn Error>> {
-        let feed_bytes = reqwest::blocking::get(&self.feed_url)?.bytes()?;
-
-        let channel = rss::Channel::read_from(&feed_bytes[..])?;
-
-        self.posts = channel
-            .items()
-            .iter()
-            .map(Post::from)
-            .filter(|post| {
-                println!("Filtering {}", post.link);
-                let week_ago =
-                    post.publication_date > Utc::now() - Duration::seconds(SECONDS_IN_WEEK);
-                let filter_words = !self.regex_filter.is_match(&post.title);
-                week_ago & filter_words
-            })
-            .collect::<Vec<Post>>();
-
-        Ok(())
-    }
-}
-
-#[derive(Debug)]
-struct Post {
-    title: String,
-    content: String,
-    summary: String,
-    publication_date: DateTime<Utc>,
-    link: String,
-}
-
-impl Post {
-    // fn summarize(&mut self) {
-    //     if self.summary.len() < 280 {
-    //         self.summary = self.content[0..280].to_string()
-    //     }
-    // }
-
-    fn summarize(&mut self) {
-        self.summary = tokio::task::block_in_place(|| {
-            tokio::runtime::Handle::current().block_on(summarize(self.content.clone()))
-        })
-        .unwrap_or_else(|_| String::from("Empty summary"));
-    }
-
-    fn to_html(&mut self) -> String {
-        self.summarize();
-        format!(
-            "<h4><a href={}>{}</a></h4><i>On {}: </i><p> {}</p>",
-            self.link,
-            self.title,
-            self.publication_date.format("%A"),
-            self.summary
-        )
-    }
-}
-
-impl From<&rss::Item> for Post {
-    fn from(item: &rss::Item) -> Self {
-        Self {
-            title: item.title().map_or_else(String::new, String::from),
-            link: item.link().map_or_else(String::new, String::from),
-            content: item.content().map_or_else(String::new, |text| {
-                let plain_text = html2text::from_read(&text.as_bytes()[..], text.len()).unwrap();
-                let text_without_urls = URL_REGEX.replace_all(plain_text.as_str(), "").to_string();
-                UNNECESSARY_SYMBOLS_REGEX
-                    .replace_all(&text_without_urls.as_str(), "")
-                    .to_string()
-            }),
-            publication_date: DateTime::parse_from_rfc2822(item.pub_date().unwrap_or_default())
-                .unwrap_or_default()
-                .with_timezone(&Utc),
-            summary: String::new(),
-        }
-    }
-}
-
-async fn post_to_telegraph(newsletter: &mut Newsletter) -> telegraph_rs::Page {
+pub async fn post_to_telegraph(newsletter: &mut Newsletter) -> telegraph_rs::Page {
     let telegraph = telegraph_rs::Telegraph::new(&newsletter.title)
         .create()
         .await
@@ -279,30 +117,4 @@ async fn post_to_telegraph(newsletter: &mut Newsletter) -> telegraph_rs::Page {
     println!("posted page");
 
     page
-}
-
-async fn summarize(text_to_summarize: String) -> Result<String, Box<dyn Error>> {
-    let summarization_prompt = String::from("Summarize the following text into a single text of no more than 280 characters. Focus on capturing the main takeaway and presenting it. Avoid excessive detail but ensure the core message is clear and engaging. Text:");
-    let prompt = format!("{} {}", summarization_prompt, text_to_summarize);
-
-    // let config = OpenAIConfig::new().with_api_base("https://api.perplexity.ai");
-    // let client = Client::with_config(config);
-    let client = Client::new();
-
-    let request = CreateChatCompletionRequestArgs::default()
-        .model("gpt-4o-mini")
-        .messages([ChatCompletionRequestSystemMessageArgs::default()
-            .content(prompt)
-            .build()?
-            .into()])
-        .max_tokens(280_u32)
-        .build()?;
-
-    let response = client.chat().create(request).await?;
-
-    response.choices[0]
-        .message
-        .content
-        .clone()
-        .ok_or("error with message option".into())
 }
