+++ digest/src/models/mod.rs
@@ -1,0 +1,5 @@
+mod feed;
+mod newsletter;
+mod post;
+
+pub use newsletter::{Newsletter, NewsletterConfig};
